revix_mcp/requirements.venv.txt
mcp>=1.0.0
motor>=3.5
pymongo>=4.7
pydantic>=2.6
python-dotenv>=1.0
httpx>=0.27
pytest>=8.0
pytest-asyncio>=0.23
